Add alternatives to the variable symbol that are easier to understand.

The variable symbol codes the status of a tree in a very succint way. This function unpacks the meaning of symbol. Be careful: this function is specifically designed for one particular dataset (from Sinharaja) and if you use it on other data sets it will likely produce wrong results (and the function may not throw an error).

add_latest_tree_status(df)

Arguments

df
A data frame with a variable named symbol.

Value

The modified data frame.

See also

Other functions to prepare data to plot repulsive tags.: add_quadrat_and_subquadrat_from_list, add_quadrat_to_df_list, add_subquad_limits, add_subquadrat, identify_subquadrat, prep_repulsive_tags

Examples

 # Add to a single dataframe
 df <- toy_list[[1]]
 head(add_latest_tree_status(df))

#>   tag     lx     ly symbol                  sym1 latest_tree_status
#> 1 150  0.817 13.239      1 Alive in 3, Dead in 4               Dead
#> 2 455 11.895 18.574     16            Alive in 4              Alive
#> 3  29 19.832  4.899     16            Alive in 4              Alive
#> 4 178  9.460  7.077     15 Alive in 2, Dead in 3               Dead
#> 5 111  6.869 11.231      0 Alive in 1, Dead in 2               Dead
#> 6 154 19.928  4.078     16            Alive in 4              Alive

 # Add to each dataframe in a list
 df_list <- toy_list
 result <- lapply(df_list, add_latest_tree_status)
 # Just show a few rows of each dataframe
 lapply(result, head)

#> $`1`
#>   tag     lx     ly symbol                  sym1 latest_tree_status
#> 1 150  0.817 13.239      1 Alive in 3, Dead in 4               Dead
#> 2 455 11.895 18.574     16            Alive in 4              Alive
#> 3  29 19.832  4.899     16            Alive in 4              Alive
#> 4 178  9.460  7.077     15 Alive in 2, Dead in 3               Dead
#> 5 111  6.869 11.231      0 Alive in 1, Dead in 2               Dead
#> 6 154 19.928  4.078     16            Alive in 4              Alive
#> 
#> $`2`
#>   tag     lx     ly symbol                  sym1 latest_tree_status
#> 1 774  2.669 17.931      1 Alive in 3, Dead in 4               Dead
#> 2 768  7.405  4.224      0 Alive in 1, Dead in 2               Dead
#> 3  31  5.370  9.601     16            Alive in 4              Alive
#> 4 265 10.344 12.736     15 Alive in 2, Dead in 3               Dead
#> 5 170 17.390 12.635     15 Alive in 2, Dead in 3               Dead
#> 6 440  3.158 14.877      1 Alive in 3, Dead in 4               Dead
#>