Help identify_subquadrat().

Help identify_subquadrat().

add_quadrat_and_subquadrat_from_list(df_list, ...)

Arguments

df_list
A list of dataframes

...
Arguments passed to add_subquadrat(). Importantly, this allows changing the size of the subquadrats.

See also

Other functions to prepare data to plot repulsive tags.: add_latest_tree_status, add_quadrat_to_df_list, add_subquad_limits, add_subquadrat, identify_subquadrat, prep_repulsive_tags

Examples

df_list <- toy_list[1:2]
# Extracting first quadrat and showing the head of each subquadrat
added <- add_quadrat_and_subquadrat_from_list(df_list)[[1]]
lapply(added, head)

#> [[1]]
#>    tag    lx    ly symbol quadrat subquadrat
#> 4  178 9.460 7.077     15       1          1
#> 7  920 5.341 2.272      1       1          1
#> 8  272 7.020 7.402     16       1          1
#> 9  266 2.614 5.359     15       1          1
#> 13 934 2.122 9.934      0       1          1
#> 21 977 0.564 0.756     15       1          1
#> 
#> [[2]]
#>    tag     lx    ly symbol quadrat subquadrat
#> 3   29 19.832 4.899     16       1          2
#> 6  154 19.928 4.078     16       1          2
#> 10 833 11.969 3.488     15       1          2
#> 15 567 16.864 6.511      0       1          2
#> 28 303 12.801 7.981      0       1          2
#> 29 780 13.023 7.048      0       1          2
#> 
#> [[3]]
#>    tag     lx     ly symbol quadrat subquadrat
#> 2  455 11.895 18.574     16       1          3
#> 11 213 11.720 14.437     15       1          3
#> 12 518 19.832 14.337     15       1          3
#> 16 113 14.726 14.222     16       1          3
#> 17 402 17.607 14.880      1       1          3
#> 18 887 16.363 11.231      1       1          3
#> 
#> [[4]]
#>    tag    lx     ly symbol quadrat subquadrat
#> 1  150 0.817 13.239      1       1          4
#> 5  111 6.869 11.231      0       1          4
#> 14  37 7.763 17.276     16       1          4
#> 31 762 5.266 19.573      1       1          4
#> 32 506 3.857 14.853     16       1          4
#> 33 913 6.562 12.056     15       1          4
#>