To a dataframe with subquadrat variable, add plot limits.

This function helps fine tune the limits of your plots. You can live withouth it, but this function gives you some extra control. The output is a modified dataframe that includes the limits that, via plot_repulsive_tags(), will be passed to the arguments xlim and ylim of ggplot2::coord_fixed().

add_subquad_limits(df_with_subquad, quad_size = 20, shrink = 0.4)

Arguments

df_with_subquad
A dataframe with the variable subquadrat that defines the 1-4 subquadrats within each quadrat.

quad_size
Size of each quadrat.

shrink
A number, generally smaller than one, giving how much to shrink the plot.

Value

A modified data frame.

Details

Plots produced with ggplot by default print with a margin around the limits set by the user. To remove that extra margin and maximize space, this function shrinks the limits a little. Be sure not to shrink so much that you loose data; The numbers on the x and y axes should include the limits that you expect. For example, if your plot is 20x20 meters, each subplot will be 10x10 meters, so your plot axes should show either the 0 to 10, or 10 to 20. If you do not read 0 and/or 10, or 10 and/or 20, your plot may not show the data you expect.

See also

Other functions to prepare data to plot repulsive tags.: add_latest_tree_status, add_quadrat_and_subquadrat_from_list, add_quadrat_to_df_list, add_subquadrat, identify_subquadrat, prep_repulsive_tags

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Showing only 1 quadrat to save space
with_subquad_list <- toy_list[1] %>%
  add_quadrat_and_subquadrat_from_list()
str(with_subquad_list)

#> List of 1
#>  $ :List of 4
#>   ..$ :'data.frame':	80 obs. of  6 variables:
#>   .. ..$ tag       : int [1:80] 178 920 272 266 934 977 875 986 59 695 ...
#>   .. ..$ lx        : num [1:80] 9.46 5.34 7.02 2.61 2.12 ...
#>   .. ..$ ly        : num [1:80] 7.08 2.27 7.4 5.36 9.93 ...
#>   .. ..$ symbol    : int [1:80] 15 1 16 15 0 15 1 1 16 16 ...
#>   .. ..$ quadrat   : chr [1:80] "1" "1" "1" "1" ...
#>   .. ..$ subquadrat: int [1:80] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ :'data.frame':	74 obs. of  6 variables:
#>   .. ..$ tag       : int [1:74] 29 154 833 567 303 780 687 284 786 372 ...
#>   .. ..$ lx        : num [1:74] 19.8 19.9 12 16.9 12.8 ...
#>   .. ..$ ly        : num [1:74] 4.9 4.08 3.49 6.51 7.98 ...
#>   .. ..$ symbol    : int [1:74] 16 16 15 0 0 0 16 15 16 0 ...
#>   .. ..$ quadrat   : chr [1:74] "1" "1" "1" "1" ...
#>   .. ..$ subquadrat: int [1:74] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ :'data.frame':	80 obs. of  6 variables:
#>   .. ..$ tag       : int [1:80] 455 213 518 113 402 887 78 525 70 681 ...
#>   .. ..$ lx        : num [1:80] 11.9 11.7 19.8 14.7 17.6 ...
#>   .. ..$ ly        : num [1:80] 18.6 14.4 14.3 14.2 14.9 ...
#>   .. ..$ symbol    : int [1:80] 16 15 15 16 1 1 15 15 16 1 ...
#>   .. ..$ quadrat   : chr [1:80] "1" "1" "1" "1" ...
#>   .. ..$ subquadrat: int [1:80] 3 3 3 3 3 3 3 3 3 3 ...
#>   ..$ :'data.frame':	66 obs. of  6 variables:
#>   .. ..$ tag       : int [1:66] 150 111 37 762 506 913 446 396 819 847 ...
#>   .. ..$ lx        : num [1:66] 0.817 6.869 7.763 5.266 3.857 ...
#>   .. ..$ ly        : num [1:66] 13.2 11.2 17.3 19.6 14.9 ...
#>   .. ..$ symbol    : int [1:66] 1 0 16 1 16 15 0 16 0 16 ...
#>   .. ..$ quadrat   : chr [1:66] "1" "1" "1" "1" ...
#>   .. ..$ subquadrat: int [1:66] 4 4 4 4 4 4 4 4 4 4 ...

# Pulling only one dataframe
with_subquad_df <- with_subquad_list[[1]][[1]]
head(with_subquad_df)

#>    tag    lx    ly symbol quadrat subquadrat
#> 4  178 9.460 7.077     15       1          1
#> 7  920 5.341 2.272      1       1          1
#> 8  272 7.020 7.402     16       1          1
#> 9  266 2.614 5.359     15       1          1
#> 13 934 2.122 9.934      0       1          1
#> 21 977 0.564 0.756     15       1          1

with_subquad_df %>%
  # The only "must be" is the variable `subquadrat`; we could remove `quadrat`
  select(-quadrat) %>%
  add_subquad_limits(quad_size = 20) %>%
  head()

#>   tag    lx    ly symbol subquadrat  x1  x2  y1  y2
#> 1 178 9.460 7.077     15          1 0.4 9.6 0.4 9.6
#> 2 920 5.341 2.272      1          1 0.4 9.6 0.4 9.6
#> 3 272 7.020 7.402     16          1 0.4 9.6 0.4 9.6
#> 4 266 2.614 5.359     15          1 0.4 9.6 0.4 9.6
#> 5 934 2.122 9.934      0          1 0.4 9.6 0.4 9.6
#> 6 977 0.564 0.756     15          1 0.4 9.6 0.4 9.6