Adds subquadrat variable to a data frame, using Shameema's code.

Adds subquadrat variable to a data frame, using Shameema's code.

add_subquadrat(df, x1 = c(0, 10, 10, 0), x2 = c(10, 20, 20, 10), y1 = c(0,
  0, 10, 10), y2 = c(10, 10, 20, 20))

Arguments

df
A data frame.

x1, x2, y1, y2
Parameters to set x and y limits of each subquadrat.

Value

A list of 4 data frames.

See also

Other functions to prepare data to plot repulsive tags.: add_latest_tree_status, add_quadrat_and_subquadrat_from_list, add_quadrat_to_df_list, add_subquad_limits, identify_subquadrat, prep_repulsive_tags

Examples

library(dplyr)
library(ggplot2)

# From a dataframe of one quadrat, get a list of 4 subquadrats.
one_quadrat <- toy_list[[1]]
head(one_quadrat)

#>   tag     lx     ly symbol
#> 1 150  0.817 13.239      1
#> 2 455 11.895 18.574     16
#> 3  29 19.832  4.899     16
#> 4 178  9.460  7.077     15
#> 5 111  6.869 11.231      0
#> 6 154 19.928  4.078     16

lapply(add_subquadrat(one_quadrat), head)

#> [[1]]
#>    tag    lx    ly symbol subquadrat
#> 4  178 9.460 7.077     15          1
#> 7  920 5.341 2.272      1          1
#> 8  272 7.020 7.402     16          1
#> 9  266 2.614 5.359     15          1
#> 13 934 2.122 9.934      0          1
#> 21 977 0.564 0.756     15          1
#> 
#> [[2]]
#>    tag     lx    ly symbol subquadrat
#> 3   29 19.832 4.899     16          2
#> 6  154 19.928 4.078     16          2
#> 10 833 11.969 3.488     15          2
#> 15 567 16.864 6.511      0          2
#> 28 303 12.801 7.981      0          2
#> 29 780 13.023 7.048      0          2
#> 
#> [[3]]
#>    tag     lx     ly symbol subquadrat
#> 2  455 11.895 18.574     16          3
#> 11 213 11.720 14.437     15          3
#> 12 518 19.832 14.337     15          3
#> 16 113 14.726 14.222     16          3
#> 17 402 17.607 14.880      1          3
#> 18 887 16.363 11.231      1          3
#> 
#> [[4]]
#>    tag    lx     ly symbol subquadrat
#> 1  150 0.817 13.239      1          4
#> 5  111 6.869 11.231      0          4
#> 14  37 7.763 17.276     16          4
#> 31 762 5.266 19.573      1          4
#> 32 506 3.857 14.853     16          4
#> 33 913 6.562 12.056     15          4
#> 

# Show one subquadrat of the default size
subquadrats_of_default_size <- add_subquadrat(one_quadrat)[[1]]
one_default_subquad <- filter(subquadrats_of_default_size, subquadrat == 1)
not_run({
  ggplot(one_default_subquad, aes(lx, ly)) + geom_point()
})

# Show one subquadrat of half the default size
subquadrats_half_size <- add_subquadrat(
  one_quadrat,
  x1 = (c(0, 10, 10, 0) / 2),
  x2 = (c(10, 20, 20, 10) / 2),
  y1 = (c(0, 0, 10, 10) / 2),
  y2 = (c(10, 10, 20, 20) / 2)
)[[1]]
one_half_sized_subquad <- filter(subquadrats_half_size, subquadrat == 1)
not_run({
  ggplot(one_half_sized_subquad, aes(lx, ly)) + geom_point()
})