A theme that is customized for map_tag().

Theme customized for map_tag(). You can further customize the theme by passing more arguments to ggplot2::theme() (via ...) or by replacing this theme completely with a new ggplot2::theme().

get_theme(panel_grid_major_colour = "black",
  panel_grid_minor_colour = "black", panel_grid_minor_linetype = "dotted",
  panel_background_fill = "white", plot_title_size = 20,
  plot_subtitle_size = 12, legend_position = "top",
  axis_ticks = element_blank(), axis_text_size = 12,
  legend_title = element_blank(), ...)

Arguments

panel_grid_major_colour
Colour of major grid lines.

panel_grid_minor_colour, panel_grid_minor_linetype
Colour and line type of minor grid lines.

panel_background_fill
Fill colour of the background of plotting area, drawn underneath plot.

plot_title_size, plot_subtitle_size
Size of the plot title and subtitle.

legend_position
The position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector).

axis_ticks
Tick marks along axes.

axis_text_size
Size of tick labels along axes.

legend_title
title of legend.

...
Other arguments passed to ggplot2::theme().

Value

A ggplot2::theme() customized for map_tag().

See also

ggplot2::theme().

Other functions to fine tune a plot of repulsive tags.: get_header, pad

Examples

ggplot2::theme()

#>  Named list()
#>  - attr(*, "class")= chr [1:2] "theme" "gg"
#>  - attr(*, "complete")= logi FALSE
#>  - attr(*, "validate")= logi TRUE

map::get_theme()

#> List of 9
#>  $ axis.text       :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi FALSE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.ticks      : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ legend.title    : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ legend.position : chr "top"
#>  $ panel.background:List of 5
#>   ..$ fill         : chr "white"
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi FALSE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ panel.grid.major:List of 6
#>   ..$ colour       : chr "black"
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi FALSE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ panel.grid.minor:List of 6
#>   ..$ colour       : chr "black"
#>   ..$ size         : NULL
#>   ..$ linetype     : chr "dotted"
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi FALSE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ plot.title      :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 20
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi FALSE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.subtitle   :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi FALSE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  - attr(*, "class")= chr [1:2] "theme" "gg"
#>  - attr(*, "complete")= logi FALSE
#>  - attr(*, "validate")= logi TRUE