From a list of dataframes output a dataframe with a useful id variable.

This function with quadrat and subquadrat variables add id.

identify_subquadrat(df_list, ...)

Arguments

df_list
A list of data frames.

...
Arguments passed to add_quadrat_and_subquadrat_from_list(). This is lets users change the size of subquadrat via add_subquadrat().

Value

A data frames with new variables quadrat, subquadrat and id, which combines the other two.

Details

The name of each element of the input list becomes a value of a new variable quadrat; and the variable subquadrat is also added, based on the values of the variables lx and ly that each dataframe should have.

See also

Other functions to prepare data to plot repulsive tags.: add_latest_tree_status, add_quadrat_and_subquadrat_from_list, add_quadrat_to_df_list, add_subquad_limits, add_subquadrat, prep_repulsive_tags

Examples

library(dplyr)
library(ggplot2)

df_list <- toy_list[1:2]
head(identify_subquadrat(df_list))

#>    id quadrat subquadrat tag    lx    ly symbol
#> 1 1-1       1          1 178 9.460 7.077     15
#> 2 1-1       1          1 920 5.341 2.272      1
#> 3 1-1       1          1 272 7.020 7.402     16
#> 4 1-1       1          1 266 2.614 5.359     15
#> 5 1-1       1          1 934 2.122 9.934      0
#> 6 1-1       1          1 977 0.564 0.756     15

df_list <- toy_list[1:2]

# Passing arguments to add_subquadrat(), to make subquadrats half the size
id <- identify_subquadrat(
 df_list,
 x1 = c(0, 10, 10, 0) / 2,
 x2 = c(10, 20, 20, 10) / 2,
 y1 = c(0, 0, 10, 10) / 2,
 y2 = c(10, 10, 20, 20) / 2
)
not_run({
  one_subquad <- filter(id, id == "16-3")
  ggplot(one_subquad, aes(lx, ly)) + geom_point()
})