Plot species from a list or vector.

Plot species from a list or vector.

lapply_plot_sp(species, cns_data, ...)

Arguments

species
A string of the species codes to plot (sp).

cns_data
Census data.

...
Arguments passed to ggplot2::geom_point().

Value

A list of plots. Usually you will want to save the list on a file ( e.g. a .pdf, .png, etc.)

Acknowledgement

Thanks to Gabriel Arellano and David Kenfack for ideas and feedback.

See also

ggplot2::geom_point(), grDevices::pdf(), grDevices::png().

Examples

not_run({
  # Ensure you have these packages
  library(ggplot2)
  library(dplyr)

  # Minimal data to show
  library(bci)
  cns_data <- bci::bci12full7
  # Filtering only 3 species for a minimal example.
  (species_selected <- unique(cns_data$sp)[1:3])
  cns_data_sub <- filter(cns_data, sp %in% species_selected)
  # Sampling only 1000 rows for a quick example
  cns_data_sub <- sample_n(cns_data_sub, 1000)

  # Selecting all species in the example dataset
  all_species <- unique(cns_data_sub$sp)
  pdf(onefile = TRUE)
  lapply_plot_sp(species = all_species, cns_data = cns_data_sub)
  dev.off()

  pdf(onefile = TRUE)
  lapply_plot_sp(
    species = all_species, cns_data = cns_data_sub,
    # Change the look of the points  passed to ?geom_point
    color = "blue", size = 3, shape = 1
  )
  dev.off()

})