Map a census plot -- empty or with elevation.

map_plot() produces a base on top of which you can add any ggplot2 layer (e.g. ggplot2::geom_point() and ggplot2::facet_wrap(); also geom_point_sp() and facet_wrap_sp()). map_elevation() and map_species() are convenient wrappers.

map_plot(data = NULL, elevation = NULL, xlim = NULL, ylim = NULL,
  theme = theme_map_sp(), line_size = 0.5, low = "#132B43",
  high = "#f70404", bins = NULL, label_elev = TRUE)

map_elevation(data = NULL, elevation = NULL, xlim = NULL, ylim = NULL,
  theme = theme_map_sp(), line_size = 0.5, low = "#132B43",
  high = "#f70404", bins = NULL, label_elev = TRUE)

map_species(data, elevation = NULL, xlim = NULL, ylim = NULL,
  theme = theme_map_sp(), drop_fill = FALSE, shape = 21,
  label_elev = TRUE, ...)

Arguments

data
Dataframe passed to ggplot2::geom_point().

elevation
A dataframe with variables gx, gy, and elev giving the elevation of the site.

xlim, ylim
A vector giving the limits of x, y axes, for example xlim = c(0, 1000), ylim = c(0, 500). Default limits should be OK -- they are set to be c(0, max), where max is the maximum value of gx or gy in the data set.

theme
An object of class theme.

line_size
A number to customize the width of the elevation lines.

low, high
Colours to represent the range between low and high elevation. Use colour names like low = "black", high = "red" or HEX colours like low = "#132B43", high = "#56B1F7" (for more colours google #132B43).

bins
A number. Setting bins creates evenly spaced contours in the range of the data. Integers

label_elev
Logical; FALSE removes the labels of elevation lines.

drop_fill
Logical; TRUE drops the fill legend.

shape
Numeric; Point shape passed to ggplot2::geom_point().

...
Arguments passed to ggplot2::geom_point().

Value

A ggplot.

See also

ggplot2::geom_point().

Examples

# NOT RUN {
library(ggplot2)

sp2 <- fgeo.tool::top(bciex::bci12s7mini, sp, 2)
bad_nms_elev <- bciex::bci_elevation
bci_elev <- dplyr::rename(bad_nms_elev, gx = x, gy = y)

# Plot

map_plot(sp2)

map_plot(sp2, elevation = bci_elev)

map_plot(sp2, elevation = bci_elev) +
  geom_point_sp(data = sp2)

# Elevation

map_elevation(NULL, bci_elev)

map_elevation(bci_elev)

map_elevation(sp2)

map_elevation(sp2) +
  geom_point_sp(data = sp2) +
  facet_wrap_sp()

# Species

map_species(sp2)

map_species(sp2, bci_elev, drop_fill = TRUE, label_elev = FALSE, size = 4) +
  facet_wrap_sp() +
  guides(color = "none")

map_species(sp2) +
  facet_grid_sp_h()

map_species(sp2) +
  facet_grid_sp_v()
# }