Pad strings.

pad provides a convenient way to create each line of the headers produced with get_header().

pad(string, total_width = 55, side = "right", pad = "_")

Arguments

string
A character vector.

total_width
Total width of the string after pasting each string in the vector of strings.

side
Side on which padding character is added (left, right or both).

pad
Single padding character (default is a space).

Value

A padded string that is convenient to produce each line of get_header()

Details

This function may output strings a little shorter or a little longer than expected based on total_width. This is a bug, and appears sometimes when the strings contain spaces. Because this bug is not too serious, and I leave it for now.

See also

Other functions to fine tune a plot of repulsive tags.: get_header, get_theme

Examples

pad(c("Hello", "world"))

#> [1] "Hello______________________ world______________________"

pad(c("Different pad", "and shorter"), total_width = 50, pad = ".")

#> [1] "Different pad............. and shorter............."

pad(c("Padding", "to the left"), total_width = 70, pad = ".", side = "left")

#> [1] "..........................Padding ..........................to the left"