Prepare a list of data frames to plot repulsive tags.

This function may be rarely necessary. Its input is a rare data structure. It is a list of dataframes, each corresponding to a 20x20 quadrat and with the variables: "tag", "lx", "ly", "subtag", "status", "symbol".

prep_repulsive_tags(df_list)

Arguments

df_list
A list of data frames.

Value

A list of dataframes, prepared for plot_repulsive_tags().

See also

Other functions to prepare data to plot repulsive tags.: add_latest_tree_status, add_limits_shrinked, add_limits, add_quadrat_and_subquadrat_from_list, add_quadrat_to_df_list, add_quadrat_to_one_df, add_subquadrat, identify_subquadrat

Examples

# The sinharaja package is local and the data set sinh_q20 is private.
# Choosing only two 20x20 quadrats
list_of_dataframes <- sinharaja::sinh_q20[1:2]
str(list_of_dataframes)

#> List of 2
#>  $ 15:'data.frame':	323 obs. of  6 variables:
#>   ..$ tag   : int [1:323] 4193 4194 4195 4196 4197 4198 4199 4200 4201 4202 ...
#>   ..$ lx    : num [1:323] 2.1 1 1.1 1.9 1.8 2 2.4 4.2 5.7 5.2 ...
#>   ..$ ly    : num [1:323] 0.5 1.6 1.8 1.2 1.5 3.2 0.7 2 3.5 4.5 ...
#>   ..$ subtag: int [1:323] 193 194 195 196 197 198 199 200 201 202 ...
#>   ..$ status: Factor w/ 5 levels "alive","broken below",..: 1 5 1 5 5 5 5 5 5 2 ...
#>   ..$ symbol: int [1:323] 16 16 16 16 16 16 16 16 16 16 ...
#>  $ 16:'data.frame':	362 obs. of  6 variables:
#>   ..$ tag   : int [1:362] 4477 4478 4479 4480 4481 4482 4483 4484 4485 4486 ...
#>   ..$ lx    : num [1:362] 4.4 4.5 0.6 0.8 1 0.3 2.8 4.6 1.1 2.6 ...
#>   ..$ ly    : num [1:362] 0.2 0.9 0.5 1.5 3 3.3 3.6 2.2 4 4.5 ...
#>   ..$ subtag: int [1:362] 477 478 479 480 481 482 483 484 485 486 ...
#>   ..$ status: Factor w/ 5 levels "alive","broken below",..: 1 5 3 3 2 2 3 1 5 5 ...
#>   ..$ symbol: int [1:362] 16 16 1 0 16 16 0 16 16 16 ...

prepared <- prep_repulsive_tags(list_of_dataframes)
str(prepared)

#> List of 8
#>  $ 15-1:'data.frame':	49 obs. of  15 variables:
#>   ..$ id                : chr [1:49] "15-1" "15-1" "15-1" "15-1" ...
#>   ..$ quadrat           : chr [1:49] "15" "15" "15" "15" ...
#>   ..$ subquadrat        : int [1:49] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ tag               : int [1:49] 4193 4194 4195 4196 4197 4198 4199 4200 4201 4202 ...
#>   ..$ lx                : num [1:49] 2.1 1 1.1 1.9 1.8 2 2.4 4.2 5.7 5.2 ...
#>   ..$ ly                : num [1:49] 0.5 1.6 1.8 1.2 1.5 3.2 0.7 2 3.5 4.5 ...
#>   ..$ subtag            : int [1:49] 193 194 195 196 197 198 199 200 201 202 ...
#>   ..$ status            : Factor w/ 5 levels "alive","broken below",..: 1 5 1 5 5 5 5 5 5 2 ...
#>   ..$ symbol            : int [1:49] 16 16 16 16 16 16 16 16 16 16 ...
#>   ..$ sym1              : chr [1:49] "Alive in 4" "Alive in 4" "Alive in 4" "Alive in 4" ...
#>   ..$ latest_tree_status: chr [1:49] "Alive" "Alive" "Alive" "Alive" ...
#>   ..$ x1                : num [1:49] 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 ...
#>   ..$ x2                : num [1:49] 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 ...
#>   ..$ y1                : num [1:49] 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 ...
#>   ..$ y2                : num [1:49] 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 ...
#>  $ 15-2:'data.frame':	92 obs. of  15 variables:
#>   ..$ id                : chr [1:92] "15-2" "15-2" "15-2" "15-2" ...
#>   ..$ quadrat           : chr [1:92] "15" "15" "15" "15" ...
#>   ..$ subquadrat        : int [1:92] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ tag               : int [1:92] 4214 4215 4216 4217 4218 4219 4220 4221 4222 4223 ...
#>   ..$ lx                : num [1:92] 10.5 10.5 10.5 12 11.5 11.6 10.5 15.4 14.1 14 ...
#>   ..$ ly                : num [1:92] 4.3 4.8 1.7 0.6 2.9 3.2 3.9 0.3 1 2.9 ...
#>   ..$ subtag            : int [1:92] 214 215 216 217 218 219 220 221 222 223 ...
#>   ..$ status            : Factor w/ 5 levels "alive","broken below",..: 3 5 1 5 3 5 3 1 5 3 ...
#>   ..$ symbol            : int [1:92] 0 16 16 16 0 16 0 16 16 0 ...
#>   ..$ sym1              : chr [1:92] "Alive in 1, Dead in 2" "Alive in 4" "Alive in 4" "Alive in 4" ...
#>   ..$ latest_tree_status: chr [1:92] "Dead" "Alive" "Alive" "Alive" ...
#>   ..$ x1                : num [1:92] 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 ...
#>   ..$ x2                : num [1:92] 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 ...
#>   ..$ y1                : num [1:92] 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 ...
#>   ..$ y2                : num [1:92] 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 ...
#>  $ 15-3:'data.frame':	108 obs. of  15 variables:
#>   ..$ id                : chr [1:108] "15-3" "15-3" "15-3" "15-3" ...
#>   ..$ quadrat           : chr [1:108] "15" "15" "15" "15" ...
#>   ..$ subquadrat        : int [1:108] 3 3 3 3 3 3 3 3 3 3 ...
#>   ..$ tag               : int [1:108] 4273 4328 4329 4330 4333 4334 4335 4336 4337 4338 ...
#>   ..$ lx                : num [1:108] 13.2 10.7 10 10.8 10.7 11.4 12.1 12 11.2 12.9 ...
#>   ..$ ly                : num [1:108] 10 12.1 11.6 11.4 14.7 13.9 13.6 13.1 11.5 10.6 ...
#>   ..$ subtag            : int [1:108] 273 328 329 330 333 334 335 336 337 338 ...
#>   ..$ status            : Factor w/ 5 levels "alive","broken below",..: 5 5 3 5 1 5 5 5 5 5 ...
#>   ..$ symbol            : int [1:108] 16 16 1 16 16 16 16 16 16 16 ...
#>   ..$ sym1              : chr [1:108] "Alive in 4" "Alive in 4" "Alive in 3, Dead in 4" "Alive in 4" ...
#>   ..$ latest_tree_status: chr [1:108] "Alive" "Alive" "Dead" "Alive" ...
#>   ..$ x1                : num [1:108] 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 ...
#>   ..$ x2                : num [1:108] 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 ...
#>   ..$ y1                : num [1:108] 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 ...
#>   ..$ y2                : num [1:108] 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 ...
#>  $ 15-4:'data.frame':	65 obs. of  15 variables:
#>   ..$ id                : chr [1:65] "15-4" "15-4" "15-4" "15-4" ...
#>   ..$ quadrat           : chr [1:65] "15" "15" "15" "15" ...
#>   ..$ subquadrat        : int [1:65] 4 4 4 4 4 4 4 4 4 4 ...
#>   ..$ tag               : int [1:65] 4293 4295 4296 4309 4309 4310 4311 4312 4313 4314 ...
#>   ..$ lx                : num [1:65] 4.7 3.4 2.4 5.2 0.2 5.9 6.5 6.6 5.8 6.7 ...
#>   ..$ ly                : num [1:65] 10.3 10.4 10 13.4 13.4 13.3 14.7 14.1 12.3 11.4 ...
#>   ..$ subtag            : int [1:65] 293 295 296 309 309 310 311 312 313 314 ...
#>   ..$ status            : Factor w/ 5 levels "alive","broken below",..: 5 5 5 5 5 1 5 5 3 1 ...
#>   ..$ symbol            : int [1:65] 16 16 16 16 16 16 16 16 15 16 ...
#>   ..$ sym1              : chr [1:65] "Alive in 4" "Alive in 4" "Alive in 4" "Alive in 4" ...
#>   ..$ latest_tree_status: chr [1:65] "Alive" "Alive" "Alive" "Alive" ...
#>   ..$ x1                : num [1:65] 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 ...
#>   ..$ x2                : num [1:65] 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 ...
#>   ..$ y1                : num [1:65] 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 ...
#>   ..$ y2                : num [1:65] 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 ...
#>  $ 16-1:'data.frame':	87 obs. of  15 variables:
#>   ..$ id                : chr [1:87] "16-1" "16-1" "16-1" "16-1" ...
#>   ..$ quadrat           : chr [1:87] "16" "16" "16" "16" ...
#>   ..$ subquadrat        : int [1:87] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ tag               : int [1:87] 4477 4478 4479 4480 4481 4482 4483 4484 4485 4486 ...
#>   ..$ lx                : num [1:87] 4.4 4.5 0.6 0.8 1 0.3 2.8 4.6 1.1 2.6 ...
#>   ..$ ly                : num [1:87] 0.2 0.9 0.5 1.5 3 3.3 3.6 2.2 4 4.5 ...
#>   ..$ subtag            : int [1:87] 477 478 479 480 481 482 483 484 485 486 ...
#>   ..$ status            : Factor w/ 5 levels "alive","broken below",..: 1 5 3 3 2 2 3 1 5 5 ...
#>   ..$ symbol            : int [1:87] 16 16 1 0 16 16 0 16 16 16 ...
#>   ..$ sym1              : chr [1:87] "Alive in 4" "Alive in 4" "Alive in 3, Dead in 4" "Alive in 1, Dead in 2" ...
#>   ..$ latest_tree_status: chr [1:87] "Alive" "Alive" "Dead" "Dead" ...
#>   ..$ x1                : num [1:87] 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 ...
#>   ..$ x2                : num [1:87] 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 ...
#>   ..$ y1                : num [1:87] 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 ...
#>   ..$ y2                : num [1:87] 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 ...
#>  $ 16-2:'data.frame':	101 obs. of  15 variables:
#>   ..$ id                : chr [1:101] "16-2" "16-2" "16-2" "16-2" ...
#>   ..$ quadrat           : chr [1:101] "16" "16" "16" "16" ...
#>   ..$ subquadrat        : int [1:101] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ tag               : int [1:101] 4512 4513 4514 4515 4518 4519 4520 4521 4522 4523 ...
#>   ..$ lx                : num [1:101] 10.5 10.3 10.6 10.5 11.2 11.8 12.1 10.1 10.4 12.3 ...
#>   ..$ ly                : num [1:101] 1.8 2.1 1.7 1.3 4.3 4.3 3.7 4 3 2.8 ...
#>   ..$ subtag            : int [1:101] 512 513 514 515 518 519 520 521 522 523 ...
#>   ..$ status            : Factor w/ 5 levels "alive","broken below",..: 5 3 5 2 5 5 3 3 3 5 ...
#>   ..$ symbol            : int [1:101] 16 0 16 16 16 16 0 0 0 16 ...
#>   ..$ sym1              : chr [1:101] "Alive in 4" "Alive in 1, Dead in 2" "Alive in 4" "Alive in 4" ...
#>   ..$ latest_tree_status: chr [1:101] "Alive" "Dead" "Alive" "Alive" ...
#>   ..$ x1                : num [1:101] 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 ...
#>   ..$ x2                : num [1:101] 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 ...
#>   ..$ y1                : num [1:101] 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 ...
#>   ..$ y2                : num [1:101] 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 ...
#>  $ 16-3:'data.frame':	89 obs. of  15 variables:
#>   ..$ id                : chr [1:89] "16-3" "16-3" "16-3" "16-3" ...
#>   ..$ quadrat           : chr [1:89] "16" "16" "16" "16" ...
#>   ..$ subquadrat        : int [1:89] 3 3 3 3 3 3 3 3 3 3 ...
#>   ..$ tag               : int [1:89] 4685 4686 4689 4690 4691 4692 4693 4694 4695 4696 ...
#>   ..$ lx                : num [1:89] 10.5 10.4 11 11.8 11.7 11.8 11.7 11.2 11.5 11.3 ...
#>   ..$ ly                : num [1:89] 14.8 13 15.2 15 15.7 15.6 15.5 14.6 14.4 14.3 ...
#>   ..$ subtag            : int [1:89] 685 686 689 690 691 692 693 694 695 696 ...
#>   ..$ status            : Factor w/ 5 levels "alive","broken below",..: 3 5 5 3 5 5 5 5 5 3 ...
#>   ..$ symbol            : int [1:89] 0 16 16 0 16 16 16 16 16 0 ...
#>   ..$ sym1              : chr [1:89] "Alive in 1, Dead in 2" "Alive in 4" "Alive in 4" "Alive in 1, Dead in 2" ...
#>   ..$ latest_tree_status: chr [1:89] "Dead" "Alive" "Alive" "Dead" ...
#>   ..$ x1                : num [1:89] 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 ...
#>   ..$ x2                : num [1:89] 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 ...
#>   ..$ y1                : num [1:89] 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 ...
#>   ..$ y2                : num [1:89] 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 ...
#>  $ 16-4:'data.frame':	72 obs. of  15 variables:
#>   ..$ id                : chr [1:72] "16-4" "16-4" "16-4" "16-4" ...
#>   ..$ quadrat           : chr [1:72] "16" "16" "16" "16" ...
#>   ..$ subquadrat        : int [1:72] 4 4 4 4 4 4 4 4 4 4 ...
#>   ..$ tag               : int [1:72] 4656 4657 4658 4659 4660 4661 4662 4663 4664 4665 ...
#>   ..$ lx                : num [1:72] 5.5 4.4 4 4.2 4 2.6 2.7 1.3 1.4 0.9 ...
#>   ..$ ly                : num [1:72] 11.3 13.4 12.7 12.2 11.9 11.2 10.3 10.5 12.7 11.3 ...
#>   ..$ subtag            : int [1:72] 656 657 658 659 660 661 662 663 664 665 ...
#>   ..$ status            : Factor w/ 5 levels "alive","broken below",..: 5 1 5 3 5 3 3 5 5 1 ...
#>   ..$ symbol            : int [1:72] 16 16 16 0 16 0 0 16 16 16 ...
#>   ..$ sym1              : chr [1:72] "Alive in 4" "Alive in 4" "Alive in 4" "Alive in 1, Dead in 2" ...
#>   ..$ latest_tree_status: chr [1:72] "Alive" "Alive" "Alive" "Dead" ...
#>   ..$ x1                : num [1:72] 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 ...
#>   ..$ x2                : num [1:72] 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 9.6 ...
#>   ..$ y1                : num [1:72] 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 10.4 ...
#>   ..$ y2                : num [1:72] 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 19.6 ...