Allow autoplotting the columns sp and elev.

Allow autoplotting the columns sp and elev.

sp_elev(sp, elev = NULL)

Arguments

sp

A ForestGEO-like dataframe with the column sp.

elev

A ForestGEO-like elevation list or its col dataframe -- with the column elev. The datasets you pass to sp and elev should come from the same forest plot. This is not compulsory but not doing so is most likely a mistake.

Value

An S3 object of class 'sp_elev'.

See also

Examples

assert_is_installed("fgeo.x")

species_from_luquillo <- fgeo.x::stem5
elevation_from_luquillo <- fgeo.x::elevation

species_elevation <- sp_elev(species_from_luquillo, elevation_from_luquillo)
inherits(species_elevation, "sp_elev")
#> [1] TRUE