Suffix tags of dead stems that map beyond the quadrat's edge.

Suffix tags of dead stems that map beyond the quadrat's edge.

suffix_edgy_dead(x, status_d, suffix, x_q = 20, y_q = x_q)

Arguments

x
A ViewFullTable.

status_d
A character string indicaing the value of the variable status that corresponds to dead stems.

suffix
A character string to suffix tree tags with.

x_q, y_q
Integer; the x and y size in meters of a quadrat.

Value

A modified version of x.

Examples

vft <- tibble::tibble(QX = 21, QY = 21, Tag = "01", Status = "dead")
vft

#> # A tibble: 1 x 4
#>      QX    QY Tag   Status
#>   <dbl> <dbl> <chr> <chr> 
#> 1  21.0  21.0 01    dead  

suffix_edgy_dead(
  x = vft,
  status_d = "dead",
  suffix = "_d",
  x_q = 20
)

#> One or more trees spillover.

#> # A tibble: 1 x 4
#>      QX    QY Tag   Status
#>   <dbl> <dbl> <chr> <chr> 
#> 1  21.0  21.0 01_d  dead