Theme for plot_base_census().

Theme for plot_base_census().

theme_tag_status(panel.grid.major = element_line(colour = "black"),
  panel.grid.minor = element_line(colour = "black", linetype = "dotted"),
  panel.background = element_rect(fill = "white"),
  plot.title = element_text(size = 20),
  plot.subtitle = element_text(size = 12),
  plot.caption = element_text(size = 8), legend.position = "top",
  legend.title = element_blank(), axis.ticks = element_blank(),
  axis.text = element_text(size = 12), ...)

Arguments

panel.grid.major
grid lines (element_line()). Specify major grid lines, or minor grid lines separately (using panel.grid.major or panel.grid.minor) or individually for each axis (using panel.grid.major.x, panel.grid.minor.x, panel.grid.major.y, panel.grid.minor.y). Y axis grid lines are horizontal and x axis grid lines are vertical. panel.grid.*.* inherits from panel.grid.* which inherits from panel.grid, which in turn inherits from line

panel.grid.minor
grid lines (element_line()). Specify major grid lines, or minor grid lines separately (using panel.grid.major or panel.grid.minor) or individually for each axis (using panel.grid.major.x, panel.grid.minor.x, panel.grid.major.y, panel.grid.minor.y). Y axis grid lines are horizontal and x axis grid lines are vertical. panel.grid.*.* inherits from panel.grid.* which inherits from panel.grid, which in turn inherits from line

panel.background
background of plotting area, drawn underneath plot (element_rect(); inherits from rect)

plot.title
plot title (text appearance) (element_text(); inherits from title) left-aligned by default

plot.subtitle
plot subtitle (text appearance) (element_text(); inherits from title) left-aligned by default

plot.caption
caption below the plot (text appearance) (element_text(); inherits from title) right-aligned by default

legend.position
the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)

legend.title
title of legend (element_text(); inherits from title)

axis.ticks
tick marks along axes (element_line()). Specify all tick marks (axis.ticks), ticks by plane (using axis.ticks.x or axis.ticks.y), or individually for each axis (using axis.ticks.x.bottom, axis.ticks.x.top, axis.ticks.y.left, axis.ticks.y.right). axis.ticks.*.* inherits from axis.ticks.* which inherits from axis.ticks, which in turn inherits from line

axis.text
tick labels along axes (element_text()). Specify all axis tick labels (axis.text), tick labels by plane (using axis.text.x or axis.text.y), or individually for each axis (using axis.text.x.bottom, axis.text.x.top, axis.text.y.left, axis.text.y.right). axis.text.*.* inherits from axis.text.* which inherits from axis.text, which in turn inherits from text

...
additional element specifications not part of base ggplot2. If supplied validate needs to be set to FALSE.

See also

Examples

class(theme_tag_status())

#> [1] "theme" "gg"