Add a quadrat variable to a dataframe based based on qx and qy coordinates.

Add a quadrat variable to a dataframe based based on qx and qy coordinates.

add_subquadrat(df, x_q, y_q = x_q, x_sq, y_sq = x_sq)

Arguments

df
A ViewFullTable dataframe.

x_q, y_q
Size in meters of a quadrat's side. For ForestGEO sites, a common value is 20.

x_sq, y_sq
Size in meters of a subquadrat's side. For ForestGEO-CTFS sites, a common value is 5.

Value

A dataframe with the additional variable subquadrat.

Examples

# NOT RUN {
# map::bci12vft_mini is included in this package
df <- map::bci12vft_mini
with_subquadrat <- add_subquadrat(df, 20, 20, 5, 5)
head(with_subquadrat[c("qx", "qy", "subquadrat")])
# }