Suffix tags of dead stems that map beyond the quadrat's edge.

Suffix tags of dead stems that map beyond the quadrat's edge.

suffix_edgy_dead(x, status_d, suffix, x_q = 20, y_q = x_q)

Arguments

x
A ForestGEO table, either a ViewFullTable or a census table.

status_d
A character string indicaing the value of the variable status that corresponds to dead stems.

suffix
A character string to suffix tree tags with.

x_q, y_q
Integer; the x and y size in meters of a quadrat.

Value

A modified version of x.

Examples

stem <- bciex::bci12s7mini
# Add x and y coordinates relative to each quadrat
with_lxly <- fgeo.tool::add_lxly(stem)

#> Gessing: plotdim = c(1000, 500)
#>   * If this guess is wrong, provide the correct argument `plotdim`

# Focus on dead stems
dead <- with_lxly[with_lxly$status == "D", ]

# Suffix only dead trees that spillover beyond `10`
spillover <- 10
relevant_vars <- c("tag", "lx", "ly")
suffixed_over10 <- suffix_edgy_dead(dead, "D", "*", x_q = spillover)

#> One or more trees spillover.

suffixed_over10[1:10, relevant_vars]

#>        tag        lx         ly
#> 12 001080* 13.200012  4.2000122
#> 26 003450* 16.599976 15.7999878
#> 30 007247* 16.300003 13.1999969
#> 31 007466* 10.800003  0.2999878
#> 38  012020  2.000000  1.8999939
#> 51  015340  2.000000  5.9000015
#> 52 015532* 10.200012 13.0000000
#> 58 016204*  9.500000 10.1000061
#> 61  016652  2.799988  4.6000061
#> 65  019314  4.700012  1.2000122

# To tag all dead trees, set spillover to 0
spillover <- 0
suffixed_all <- suffix_edgy_dead(dead, "D", "*", x_q = spillover)

#> One or more trees spillover.

suffixed_all[1:10, relevant_vars]

#>        tag        lx         ly
#> 12 001080* 13.200012  4.2000122
#> 26 003450* 16.599976 15.7999878
#> 30 007247* 16.300003 13.1999969
#> 31 007466* 10.800003  0.2999878
#> 38 012020*  2.000000  1.8999939
#> 51 015340*  2.000000  5.9000015
#> 52 015532* 10.200012 13.0000000
#> 58 016204*  9.500000 10.1000061
#> 61 016652*  2.799988  4.6000061
#> 65 019314*  4.700012  1.2000122